requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
requests
beautifulsoup4
streamlit
